% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\docType{data}
\name{Datos}
\alias{Datos}
\title{The following data refers to the height of a plant X1
(in m.), its root length X2 (in cm), its leaf area X3 (in cm2) and the weight of the
fruit pulp X4 (in gm.), of a variety of apple tree.}
\format{
\subsection{\code{Datos}}{

A data frame with 10 rows and 4 columns:
\describe{
\item{X1}{Plant height}
\item{X2}{Root length}
\item{X3}{Leaf area}
\item{X4}{Fruit pulp}
}
}
}
\source{
Data taken from the book "Statistical Analysis of Multivariate Data" by Professor Luis Guillermo Diaz Monroy (UNAL)
}
\usage{
Datos
}
\description{
This data set describes the study components of an apple plant.
}
\keyword{datasets}
