% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deuclidiana.R
\name{Deuclidiana}
\alias{Deuclidiana}
\title{Euclidean distance}
\usage{
Deuclidiana(Datos)
}
\arguments{
\item{Datos}{Data set}
}
\value{
Shows the Euclidean distances of the observations with respect to
the vector of means.
}
\description{
Euclidean distance
}
\examples{
Deuclidiana(Datos)

}
